backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
greenlet
google-generativeai
pydantic
python-dotenv
httpx
